Integrate Room!
This is the 4th and hopefully last (for awhile) storage implementation I've tried. The formers have been Store, the smart calladapter, Realm (never checked in), and now this. I'm pretty happy with how this turned out.

The initial Realm implementation in this work set some good separation and groundwork around how to structure this. The result is that we have a (temporary) new StorageBackedNewsController for incremental migration and simple adjustments to make an existing controller extend from it.

TODOs on this are to ultimately get rid of having two CatchUpItems and possibly merge StorageBackedNewsController into the base news controller XOR pull out its loading functionality into something more pluggable and composable (i.e. have a "news loader" and possibly wrap it in a "storage-first loader". It's also not clear where things that already have their own baked in caching like Firebase and Apollo fit into this, or if they should just force freshness and let this handle it all for consistency.