Move GithubController fully over to storage backed
This drops apollo's built-in cache resolution because it frankly doesn't seem to *work*. Some parts are cached, but the resolver is never hit in the request. This also fully moves over to the storage backed approach with Room.

This also lays some groundwork for paging, but currently doesn't do anything. Need to figure out a better way to handle paging in services that don't have numerical pages (like github and reddit's "after" notion)