Introduce Service
This is the higher level construct of what represents a service going forward. Services will be able to device themselves within this and allow for us to create them on the DI graph and injectable to generic containers. Big wins for modularity, testing, etc

For now - visual and text services are separate. May combine these in the future