Introduce NewsServiceController template
This is a template for handling Services

These are injected via multibinding that contains providers for all services collected. The key is passed in as an argument to the controller. This is a slightly tweaked version of the service controllers of before that operates entirely based on a given service