Implement StorageBackedService, wire up in Dagger
This is a reimplementation of StorageBackedNewsController in the new Service system. It's smarter, String-page supporting (but numeric fast track still works!), and more robust. Learned a lot in writing this. Made a few API tweaks to Service/meta along the way. Main way this works is that services must return a DataResult with an optional next page token