Implement HintArbiter
HintArbiter is a manager for hints and powered by TapTargetView. Hints can be enqueued and they'll be shown in a sequence with other hints in a given window. This is done by leveraging a debounced buffer mechanism under the hood. Hints will be shown after no other requests have come within a 2 second window.

Note that TapTargetView is not included in this yet because other build.gradle changes. Sorry git bisect :X