Implement adapter caching
This creates adapter properties rather than always looking up on a saved moshi property. The properties are kept by their TypeName and not necessarily 1:1 with properties, so we can reuse/share adapters for the same types. Names are simplified to be more human readable (from auto-value-gson).

This also includes a few opportunistic cleanups:
- Use a NameAllocator for all names via allocate() extension fun
- Removed a stale debugging check