Upgrade to gradle 5.1-rc-1
have to upgrade apollo to 1.0.1-SNAPSHOT avoid this problem:
https://github.com/apollographql/apollo-android/issues/1135

Apollo 1.0.1 has a new api for customTypeMapping

Gradle 5.1 requires explicit dep on soLoader
With 5.1, the wrapper task conflicts with built in task.

To execute the wrapper task you can just do:
`./gradlew wrapper --gradle-version 5.1-rc-1 --distribution-type all --no-daemon`

Run it twice if you want the scripts and jar to be updated.