Upload processed lint sarif (#479)
This startups uploading lint `.sarif` outputs to GitHub. This is
something we more or less get for free with android lint and can upload
via github's upload-sarif action.

This takes the extra step though to _filter_ those results first,
because ultimately most lint report contents are _not_ relevant to the
PR's changed files. I'm not totally sure if this is the best way to do
it, it would be nice if we could also diff them with the existing lint
(probably via baseline) and instead only show _new_ lints (and include
those in files not in the PR, since lints like `@RestrictTo` could
affect that).

This at least works for catchup's purposes though.

Other fixes
- Fixed release apks not getting built on CI
- Fix lintRelease not being run on CI
- Fix release APK build