Revamp LumberYard (#1159)
This revamps `LumberYard` and scratches some other itches along the way

There are two implementations of LumberYard - one simple one with an
in-memory ring buffer, and one comprehensive disk-writing implementation
with coroutines. CatchUp is fine with just the simple buffer and uses
that.

- `DiskLumberYard`
  - Not really necessary for CatchUp's needs but a fun implementation.
- Pending, unflushed logs are buffered into a `Channel` and periodically
flushed to files on disk. These files are rotated once they reach 1MB in
size. Currently the files are unused, but could be attached to the bug
reporter API if github ever supported file attachmenets.
- Last 200 logs are buffered in-memory in a ringbuffer to attach as a
tab in bugsnag reports or viewing in the debug UI.
- Optionally an `AtomicFile` for file writes. Not really necessary given
the appending nature though.
- Yolo/best-effort sends fatal exceptions to it when the app is crashing
+ attempts to flush the buffer
  - Has tests!
- Open question for another day – What's the right behavior for flushing
when the app is foregrounded/backgrounded?
- Implements an `Okio`-based (and mostly KMP-friendly) `AtomicFile` +
tests